refactor(account-set): isolate set validation
Set batch validation is a pure graph decision, but it lived inside the
cache adapter beside snapshot and SQL fallback concerns. That coupling
made topology invariants harder to review and left the core algorithm
without infrastructure-free tests.

Move cycle, path-uniqueness, account-reachability, and depth decisions
into a leaf graph_validation module. Keep cache adaptation and
repository IO at their boundaries, and add focused tests for existing
and proposed edge interactions.
